+++ packages/orq-rc/RELEASES.md
@@ -1439,3 +1439,13 @@ Based on:
 - [typescript v3.4.0-rc.31] packages/orq-rc
 ### Releases
 - [NPM v3.4.0-rc.31] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.31 - packages/orq-rc
+
+## 2025-04-16 18:18:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.533.0 (2.578.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.4.0-rc.32] packages/orq-rc
+### Releases
+- [NPM v3.4.0-rc.32] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.32 - packages/orq-rc